ai_memory_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
nltk==3.8.1
scipy==1.11.4
pydantic==2.5.0 